Change get_info_int to use strtoll instead of strtol

+++ include/wb_tools.h
@@ -36,10 +36,10 @@ char *get_info_first(const char *input,
                      const char *patt_e_list,
                      const char *desc);
 
-long int get_info_int(const char *input,
-                      const char *patt_b,
-                      const char *patt_e,
-                      const char *desc);
+long long int get_info_int(const char *input,
+                           const char *patt_b,
+                           const char *patt_e,
+                           const char *desc);
 
 char *base64encode(const void *input, size_t inlength);
 char *base64decode(const void *input, size_t inlength, size_t *outlength);
